# app/static/swagger.yaml
swagger: '2.0'

info:
  title: "Mechanic Shop API"
  description: |
    The Mechanic Shop API allows a repair shop to manage customers, mechanics,
    service tickets, and parts inventory.

    NOTE:
      - Mechanics do not have login functionality in this system.

    Authentication:
    Protected routes require a Bearer token:
    Authorization: Bearer <token>
  version: "1.0.0"

host: "127.0.0.1:5000"
schemes:
  - "http"

consumes:
  - "application/json"
produces:
  - "application/json"

securityDefinitions:
  bearerAuth:
    type: apiKey
    name: authorization
    in: header

paths:

  /customers/login:
    post:
      tags: [Customers]
      summary: "Login"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"

  /customers:
    post:
      tags: [Customers]
      summary: "Create customer"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CustomerPayload"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/CustomerResponse"

    get:
      tags: [Customers]
      summary: "Get all customers"
      parameters:
        - in: query
          name: page
          type: integer
        - in: query
          name: per_page
          type: integer
      responses:
        200:
          description: "Customer list"
          schema:
            $ref: "#/definitions/AllCustomers"

  /customers/{customer_id}:
    put:
      tags: [Customers]
      summary: "Update customer"
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: customer_id
          type: integer
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          description: "Updated"
          schema:
            $ref: "#/definitions/CustomerResponse"

    delete:
      tags: [Customers]
      summary: "Delete customer"
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: customer_id
          type: integer
          required: true
      responses:
        200:
          description: "Deleted"
          schema:
            $ref: "#/definitions/DeleteCustomerResponse"

  /customers/my-tickets:
    get:
      tags: [Customers]
      summary: "Get my tickets"
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Tickets"
          schema:
            $ref: "#/definitions/AllServiceTickets"

  /mechanics:
    get:
      tags: [Mechanics]
      summary: "Get mechanics"
      responses:
        200:
          schema:
            $ref: "#/definitions/AllMechanics"

    post:
      tags: [Mechanics]
      summary: "Create mechanic"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CreateMechanicPayload"
      responses:
        201:
          schema:
            $ref: "#/definitions/MechanicResponse"

  /mechanics/{mechanic_id}:
    get:
      tags: [Mechanics]
      summary: "Get mechanic"
      parameters:
        - in: path
          name: mechanic_id
          type: integer
          required: true
      responses:
        200:
          schema:
            $ref: "#/definitions/MechanicResponse"

    put:
      tags: [Mechanics]
      summary: "Update mechanic"
      parameters:
        - in: path
          name: mechanic_id
          type: integer
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/UpdateMechanicPayload"
      responses:
        200:
          schema:
            $ref: "#/definitions/MechanicResponse"

    delete:
      tags: [Mechanics]
      summary: "Delete mechanic"
      parameters:
        - in: path
          name: mechanic_id
          type: integer
          required: true
      responses:
        200:
          schema:
            $ref: "#/definitions/DeleteMechanicResponse"

  /mechanics/most_tickets:
    get:
      tags: [Mechanics]
      summary: "Mechanics sorted by tickets"
      responses:
        200:
          schema:
            $ref: "#/definitions/AllMechanics"

  /mechanics/search:
    get:
      tags: [Mechanics]
      summary: "Search mechanics"
      parameters:
        - in: query
          name: name
          type: string
          required: true
      responses:
        200:
          schema:
            $ref: "#/definitions/AllMechanics"

  /inventories:
    get:
      tags: [Inventory]
      summary: "Get inventory"
      responses:
        200:
          schema:
            $ref: "#/definitions/AllInventory"

    post:
      tags: [Inventory]
      summary: "Create part"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CreateInventoryPayload"
      responses:
        201:
          schema:
            $ref: "#/definitions/InventoryResponse"

  /inventories/{part_id}:
    put:
      tags: [Inventory]
      summary: "Update part"
      parameters:
        - in: path
          name: part_id
          type: integer
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/UpdateInventoryPayload"
      responses:
        200:
          schema:
            $ref: "#/definitions/InventoryResponse"

    delete:
      tags: [Inventory]
      summary: "Delete part"
      parameters:
        - in: path
          name: part_id
          type: integer
          required: true
      responses:
        200:
          schema:
            $ref: "#/definitions/DeleteInventoryResponse"

  /service-tickets:
    get:
      tags: [ServiceTickets]
      summary: "Get all service tickets"
      responses:
        200:
          description: List of all service tickets
          schema:
            $ref: "#/definitions/AllServiceTickets"

    post:
      tags: [ServiceTickets]
      summary: "Create a new service ticket"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CreateServiceTicketPayload"
      responses:
        201:
          description: Ticket successfully created
          schema:
            $ref: "#/definitions/ServiceTicketResponse"
        400:
          description: Validation error

  /{service_ticket_id}:
    put:
      tags: [ServiceTickets]
      summary: "Update a service ticket"
      parameters:
        - in: path
          name: service_ticket_id
          type: integer
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/EditServiceTicketPayload"
      responses:
        200:
          description: Updated ticket
          schema:
            $ref: "#/definitions/ServiceTicketResponse"
        404:
          description: Ticket not found
        400:
          description: Validation error

  /{ticket_id}/assign-mechanic/{mechanic_id}:
    put:
      tags: [ServiceTickets]
      summary: "Assign a mechanic to a ticket"
      parameters:
        - in: path
          name: ticket_id
          type: integer
          required: true
        - in: path
          name: mechanic_id
          type: integer
          required: true
      responses:
        200:
          description: Mechanic assigned
          schema:
            $ref: "#/definitions/ServiceTicketResponse"
        404:
          description: Ticket or mechanic not found
        400:
          description: Mechanic already assigned

  /{ticket_id}/remove-mechanic/{mechanic_id}:
    put:
      tags: [ServiceTickets]
      summary: "Remove a mechanic from a ticket"
      parameters:
        - in: path
          name: ticket_id
          type: integer
          required: true
        - in: path
          name: mechanic_id
          type: integer
          required: true
      responses:
        200:
          description: Mechanic removed
          schema:
            $ref: "#/definitions/ServiceTicketResponse"
        404:
          description: Ticket or mechanic not found
        400:
          description: Mechanic not assigned

  /{ticket_id}/edit:
    put:
      tags: [ServiceTickets]
      summary: "Bulk add/remove mechanics from a ticket"
      parameters:
        - in: path
          name: ticket_id
          type: integer
          required: true
        - in: body
          name: body
          schema:
            $ref: "#/definitions/EditMechanicsPayload"
      responses:
        200:
          description: Updated ticket mechanics
          schema:
            $ref: "#/definitions/ServiceTicketResponse"
        404:
          description: Ticket not found
        400:
          description: Validation error

  /{ticket_id}/add-part/{part_id}:
    put:
      tags: [ServiceTickets]
      summary: "Add a part to a service ticket"
      parameters:
        - in: path
          name: ticket_id
          type: integer
          required: true
        - in: path
          name: part_id
          type: integer
          required: true
      responses:
        200:
          description: Part added
          schema:
            $ref: "#/definitions/ServiceTicketResponse"
        404:
          description: Ticket or part not found
        400:
          description: Part already added

definitions:

  LoginCredentials:
    type: object
    required: [email, password]
    properties:
      email:
        type: string
        format: email
      password:
        type: string

  LoginResponse:
    type: object
    properties:
      token:
        type: string
      message:
        type: string

  CustomerPayload:
    type: object
    required: [name, email, phone, password]
    properties:
      name:
        type: string
      email:
        type: string
        format: email
      phone:
        type: string
      password:
        type: string

  UpdateCustomerPayload:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      password:
        type: string

  CustomerResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      phone:
        type: string

  AllCustomers:
    type: array
    items:
      $ref: "#/definitions/CustomerResponse"

  DeleteCustomerResponse:
    type: object
    properties:
      message:
        type: string

  CreateMechanicPayload:
    type: object
    required: [name, email, phone, salary]
    properties:
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      salary:
        type: number

  UpdateMechanicPayload:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      salary:
        type: number

  MechanicResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      salary:
        type: number

  AllMechanics:
    type: array
    items:
      $ref: "#/definitions/MechanicResponse"

  DeleteMechanicResponse:
    type: object
    properties:
      message:
        type: string

  CreateInventoryPayload:
    type: object
    required: [part_name, price]
    properties:
      part_name:
        type: string
      price:
        type: number

  UpdateInventoryPayload:
    type: object
    properties:
      part_name:
        type: string
      price:
        type: number

  InventoryResponse:
    type: object
    properties:
      id:
        type: integer
      part_name:
        type: string
      price:
        type: number

  AllInventory:
    type: array
    items:
      $ref: "#/definitions/InventoryResponse"

  DeleteInventoryResponse:
    type: object
    properties:
      message:
        type: string

  ServiceTicketResponse:
    type: object
    properties:
      id:
        type: integer
      vin:
        type: string
      service_desc:
        type: string
      service_date:
        type: string
        format: date-time
      customer_id:
        type: integer
      mechanics:
        type: array
        items:
          type: integer
        description: List of mechanic IDs assigned to the ticket
      inventory:
        type: array
        items:
          type: integer
        description: List of part IDs assigned to the ticket


  AllServiceTickets:
    type: array
    items:
      $ref: "#/definitions/ServiceTicketResponse"

  CreateServiceTicketPayload:
    type: object
    required:
      - vin
      - service_desc
      - customer_id
    properties:
      vin:
        type: string
      service_desc:
        type: string
      customer_id:
        type: integer
      service_date:
        type: string
        format: date-time

  EditServiceTicketPayload:
    type: object
    properties:
      vin:
        type: string
      service_desc:
        type: string
      customer_id:
        type: integer
      service_date:
        type: string
        format: date-time

  EditMechanicsPayload:
    type: object
    properties:
      add_ids:
        type: array
        items:
          type: integer
      remove_ids:
        type: array
        items:
          type: integer